feat(gpu): add an entry point to decompress integer server key to cuda server key
